Destroying x64 Warnings :no_pedestrians:
Changed uint sizes to size_t
Added template Push() function to RE_Json
I'm scared to tackle warnings in RE_Shader :worried:

+++ RedEye/Engine/RE_CompCamera.h
@@ -92,7 +92,7 @@ public:
 	eastl::vector<const char*> GetAllResources() override;
 
 	// Serialization
-	unsigned int GetBinarySize() const override;
+	size_t GetBinarySize() const override;
 	void SerializeBinary(char*& cursor, eastl::map<const char*, int>* resources) const override;
 	void DeserializeBinary(char*& cursor, eastl::map<int, const char*>* resources)override;
 	void SerializeJson(RE_Json* node, eastl::map<const char*, int>* resources) const override;
